+++ lib/presentation/pages/quran_library_screen.dart
@@ -3,7 +3,6 @@ import 'dart:developer';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:fluttertoast/fluttertoast.dart';
 import 'package:get/get.dart';
 import 'package:quran_library/core/extensions/convert_number_extension.dart';
 import 'package:quran_library/core/extensions/extensions.dart';